seperate sync out of create-db
