+++ create-db/entrypoint.sh
@@ -1,17 +1,7 @@
 #!/bin/bash
 set -e
 
-echo "::group::Preparing files"
 init_path=$PWD
-mkdir -p /tmp/repo/
-mkdir upload_packages
-# do not recursively find files
-find ${package_path} -type f -path "*.tar.zst" -exec cp {} ./upload_packages/ \;
-if [[ -d "/tmp/repo/old" ]]; then
-    rm -rf /tmp/repo/old
-fi
-cp -r $repo_path  /tmp/repo/old
-echo "::endgroup::"
 
 echo "::group::Importing GPG key"
 if [ ! -z "$gpg_key" ]; then
@@ -19,21 +9,4 @@ if [ ! -z "$gpg_key" ]; then
 fi
 echo "::endgroup::"
 
-pushd upload_packages || exit 1
-
-python3 $init_path/create-db/create-db.py
-
-popd
-
-if [[ -d "$output_path" ]]; then
-    rm -rf $output_path/*
-else
-    mkdir -p $output_path
-fi
-
-cp -r upload_packages/* $output_path/
-
-echo "::group::Cleaning up"
-rm -rf /tmp/repo
-rm -rf upload_packages
-echo "::endgroup::"
+python3 $init_path/create-db/create_db.py
