+++ sync-db/entrypoint.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+
+init_path=$PWD
+python3 $init_path/sync-db/sync_database.py
